requirements.txt
h5py==3.1.0
ortools==8.0.8283
pyyaml==5.4.1
numpy==1.19.5
torch==1.8.1+cu111 torchvision==0.9.1+cu111 torchaudio===0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
torch-scatter -f https://pytorch-geometric.com/whl/torch-1.7.0+cu110.html
torch-sparse>=0.4.3
torch-cluster>=1.4.5
torch-geometric>=1.3.2
